requirements.txt
django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
requests>=2.31
psycopg2-binary>=2.9
Pillow>=10.0
python-dotenv>=1.0
gunicorn>=21.2
